CB-3613 refactor and move gateway public endpoint (cert/dns entry) managing logic after host metadata collection state at creation
